Create separate llm provider directory, add response_schema to ollama provider
