+++ timesketch/api/v1/resources/llm_summarize.py
@@ -28,7 +28,8 @@ from flask_login import login_required, current_user
 from flask_restful import Resource
 
 from timesketch.api.v1 import resources, export
-from timesketch.lib import definitions, llms, utils
+from timesketch.lib import definitions, utils
+from timesketch.lib.llms.providers import manager
 from timesketch.lib.definitions import METRICS_NAMESPACE
 from timesketch.models.sketch import Sketch
 
@@ -304,8 +305,8 @@ class LLMSummarizeResource(resources.ResourceMixin, Resource):
                 configured LLM provider
         """
         try:
-            feature_name = "llm_summarization"
-            llm = llms.manager.LLMManager.create_provider(feature_name=feature_name)
+            feature_name = "llm_summarize"
+            llm = manager.LLMManager.create_provider(feature_name=feature_name)
         except Exception as e:  # pylint: disable=broad-except
             logger.error("Error LLM Provider: %s", e)
             abort(
