+++ timesketch/lib/llms/providers/__init__.py
@@ -11,8 +11,8 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""LLM module for Timesketch."""
+"""LLM providers for Timesketch."""
 
-from timesketch.lib.llms import ollama
-from timesketch.lib.llms import vertexai
-from timesketch.lib.llms import aistudio
+from timesketch.lib.llms.providers import ollama
+from timesketch.lib.llms.providers import vertexai
+from timesketch.lib.llms.providers import aistudio
