+++ timesketch/lib/llms/providers/aistudio.py
@@ -1,4 +1,4 @@
-# Copyright 2024 Google Inc. All rights reserved.
+# Copyright 2025 Google Inc. All rights reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -15,8 +15,8 @@
 
 import json
 from typing import Optional
-from timesketch.lib.llms import interface
-from timesketch.lib.llms import manager
+from timesketch.lib.llms.providers import interface
+from timesketch.lib.llms.providers import manager
 
 
 # Check if the required dependencies are installed.
